feat: Move result logic into Dealer
