+++ src/main/kotlin/blackjack/domain/Dealer.kt
@@ -1,5 +1,7 @@
 package blackjack.domain
 
+import blackjack.view.dto.PlayerGameResult
+
 private val DEALER_MIN_POINT = Point(17)
 
 class Dealer(playingCards: PlayingCards = PlayingCards()) {
@@ -27,4 +29,9 @@ class Dealer(playingCards: PlayingCards = PlayingCards()) {
     }
 
     fun firstCard(): Set<Card> = player.firstCard()
+    fun result(players: List<Player>): List<PlayerGameResult> {
+        val playerGameResults = players.map { PlayerGameResult(it, it.flip(this)) }
+        val dealerBettingAmount = -playerGameResults.sumOf { it.winningAmount }
+        return listOf(PlayerGameResult(this, dealerBettingAmount)) + playerGameResults
+    }
 }
